requirements.txt
# Core dependencies for Docker deployment
pandas==2.2.3
numpy==1.26.4
pydantic==2.9.2
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
plotly==5.24.1
sqlalchemy>=1.4,<2.0.36
typing-extensions==4.12.2

# Optional: LangGraph dependencies (uncomment if you have OpenAI API key)
# langgraph==0.2.45
# langchain==0.3.7
# langchain-openai==0.2.8
# langchain-community==0.3.7
